Add methods to AdminRestController. Refactor project

+++ src/main/java/ru/internship/voting/repository/DishRepository.java
@@ -11,25 +11,36 @@ import java.util.List;
 public class DishRepository {
 
     private final DataJpaDishRepository jpaDishRepository;
+    private final DataJpaRestaurantRepository jpaRestaurantRepository;
 
     @Autowired
-    public DishRepository(DataJpaDishRepository jpaDishRepository) {
+    public DishRepository(DataJpaDishRepository jpaDishRepository, DataJpaRestaurantRepository jpaRestaurantRepository) {
         this.jpaDishRepository = jpaDishRepository;
+        this.jpaRestaurantRepository = jpaRestaurantRepository;
     }
 
-    public Dish get(int id) {
-        return jpaDishRepository.findById(id).orElse(null);
+    public Dish get(int id, int restId) {
+        return jpaDishRepository.findById(id)
+                .filter(dish -> dish.getRestaurant().getId() == restId)
+                .orElse(null);
     }
 
     public List<Dish> getAll(int restaurantId) {
         return jpaDishRepository.getAll(restaurantId);
     }
 
-    public boolean delete(int id) {
+    public boolean delete(int id, int restId) {
+        if (get(id, restId) == null) {
+            return false;
+        }
         return jpaDishRepository.delete(id) != 0;
     }
 
-    public Dish save(Dish dish) {
+    public Dish save(Dish dish, int restId) {
+        if (!dish.isNew() && get(dish.getId(), restId) == null) {
+            return null;
+        }
+        dish.setRestaurant(jpaRestaurantRepository.getOne(restId));
         return jpaDishRepository.save(dish);
     }
 
@@ -40,5 +51,4 @@ public class DishRepository {
     public Dish getWithRestaurant(int id) {
         return jpaDishRepository.getWithRestaurant(id);
     }
-
 }
